# openapi.yaml
openapi: 3.1.0
info:
  title: Additizer API
  version: 1.0.0
  description: REST API for Additizer authentication, users, preset groups, and presets.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/groups/{groupID}/presets:
    parameters:
      - $ref: "#/components/parameters/GroupID"
    get:
      tags:
        - Presets
      operationId: listPresetsInGroup
      summary: List presets in a group
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Presets in the selected group, including preset JSON data.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PresetWithData"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/presets/:
    get:
      tags:
        - Presets
      operationId: listPresets
      summary: List presets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Presets visible to the authenticated user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Preset"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Presets
      operationId: createPreset
      summary: Create a preset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePresetRequest"
      responses:
        "201":
          description: Preset created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preset"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/presets/{presetID}:
    parameters:
      - $ref: "#/components/parameters/PresetID"
    put:
      tags:
        - Presets
      operationId: updatePreset
      summary: Update a preset
      description: When either appVersion or preset is provided, both must be provided together.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePresetRequest"
      responses:
        "200":
          description: Preset updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preset"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Presets
      operationId: deletePreset
      summary: Delete a preset
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Preset deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/presets/{presetID}/share:
    parameters:
      - $ref: "#/components/parameters/PresetID"
    post:
      tags:
        - Presets
      operationId: sharePreset
      summary: Share a preset with another user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareWithUserRequest"
      responses:
        "201":
          description: Share record created or restored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GroupID:
      name: groupID
      in: path
      required: true
      schema:
        type: integer
        format: uint
        minimum: 1
      description: Preset group ID.
    PresetID:
      name: presetID
      in: path
      required: true
      schema:
        type: integer
        format: uint
        minimum: 1
      description: Preset ID.
  responses:
    BadRequest:
      description: Invalid JSON, path parameter, or validation failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing, invalid, expired, or insufficient authentication.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "validation failed: name is required"
...
    ModuleType:
      type: string
      enum:
        - harmonics-editor
        - oscillator
        - spectral-filter
        - envelope
        - spectral-mixer
        - spectral-blend
        - mixer
        - amplifier
        - waveshaper
    PresetData:
      description: Arbitrary valid JSON preset payload.
    CreatePresetRequest:
      type: object
      required:
        - groupId
        - type
        - name
        - public
        - appVersion
        - preset
      additionalProperties: false
      properties:
        groupId:
          type: integer
          format: uint
          minimum: 1
        type:
          $ref: "#/components/schemas/ModuleType"
        name:
          type: string
          minLength: 1
          maxLength: 255
        public:
          type: boolean
        appVersion:
          type: string
          description: Strict semantic version.
          maxLength: 255
          examples:
            - 1.0.0
        preset:
          $ref: "#/components/schemas/PresetData"
    UpdatePresetRequest:
      type: object
      required:
        - type
        - name
        - public
      additionalProperties: false
      properties:
        type:
          $ref: "#/components/schemas/ModuleType"
        name:
          type: string
          minLength: 1
          maxLength: 255
        public:
          type: boolean
        appVersion:
          type: string
          description: Strict semantic version. Required when preset is provided.
          maxLength: 255
          examples:
            - 1.0.1
        preset:
          allOf:
            - $ref: "#/components/schemas/PresetData"
          description: Required when appVersion is provided.
    Preset:
      type: object
      required:
        - id
        - groupId
        - type
        - name
        - public
        - appVersion
      properties:
        id:
          type: integer
          format: uint
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        groupId:
          type: integer
          format: uint
          minimum: 1
        groupName:
          type: string
        type:
          $ref: "#/components/schemas/ModuleType"
        name:
          type: string
        public:
          type: boolean
        appVersion:
          type: string
        preset:
          $ref: "#/components/schemas/PresetData"
    PresetWithData:
      allOf:
        - $ref: "#/components/schemas/Preset"
        - type: object
          required:
            - preset
          properties:
            preset:
              $ref: "#/components/schemas/PresetData"
...
    ShareWithUserRequest:
      type: object
      required:
        - shareWithUserId
      additionalProperties: false
      properties:
        shareWithUserId:
          type: string
          format: uuid
    ShareRecord:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: uint
          minimum: 1
          description: Preset share or preset group share row id.